Move model-related submodules to model module
